+++ NIMIO2024_TFG_PolMorenoCastelar/src/main/java/com/nimio2024/nimio2024_tfg_polmorenocastelar/application/CenterRestController.java
@@ -3,6 +3,7 @@ package com.nimio2024.nimio2024_tfg_polmorenocastelar.application;
 import com.nimio2024.nimio2024_tfg_polmorenocastelar.application.dto.CenterDTO;
 import com.nimio2024.nimio2024_tfg_polmorenocastelar.controllers.CenterController;
 import com.nimio2024.nimio2024_tfg_polmorenocastelar.domain.Center;
+import com.nimio2024.nimio2024_tfg_polmorenocastelar.domain.School;
 import com.nimio2024.nimio2024_tfg_polmorenocastelar.service.CenterService;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
@@ -20,6 +21,11 @@ public class CenterRestController {
     @GetMapping("")
     public ResponseEntity<List<Center>> getAllCenters() {
         return centerController.getAllCenters();
+    }
+
+    @GetMapping("/{centerId}/schools")
+    public List<School> getSchoolsByCenter(@PathVariable Long centerId) {
+        return centerController.getSchoolsByCenter(centerId);
     }
 
     @PostMapping("")